+++ packages/melody-compiler/__tests__/IfSpec.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { parse } from 'melody-parser';
+import { parse, getNodeSource } from 'melody-parser';
 import { extension } from 'melody-extension-core';
 
 describe('if', function() {
@@ -57,5 +57,18 @@ describe('if', function() {
             extension
         );
         expect(node.expressions[0]).toMatchSnapshot();
+    });
+
+    test('should be able to reproduce the original source', function() {
+        const ifStatementSource = `{% if foo %}
+        Foo
+    {%- else -%}
+        Bar
+    {% endif %}`;
+        const source = `The winner is: ${ifStatementSource}`;
+        const node = parse(source, extension);
+        expect(getNodeSource(node.expressions[1], source)).toEqual(
+            ifStatementSource
+        );
     });
 });