+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ### Features
 
 -   `[melody-parser]`: Introduce option `preserveSourceLiterally`
+-   `[melody-parser]`: Add parsing of declarations like `<!DOCTYPE html>`, along with new Parser option `ignoreDeclarations`.
+-   `[melody-parser]`: Fix bug where `1..5` results in a `MemberExpression` instead of a `BinaryRangeExpression`
+
+### Bug fixes
+
+-   `[melody-parser]`: Add previously missing `trimLeft` and `trimRight` properties to `PrintExpressionStatement` nodes
 
 ## 1.5.0